+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # MountsJournal
 
-* Added option "Enable arrow buttons to browse mounts".
-* Test: map events removed.
+* Updated the Selected filter.